PY/Bots/AOEUbot/requirements.txt
Pillow==6.0.0
Pillow-SIMD==5.3.0.post1
PyAutoGUI==0.9.42
pytesseract==0.2.6
